Commit 3: Authentication

+++ backend/controllers/goalController.js
@@ -1,12 +1,15 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc   Get goals 
 // @route  Get /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({
+        user: req.user.id
+    })
     res.status(200).json(goals)
 })
 
@@ -20,7 +23,8 @@ const setGoals = asyncHandler(async (req, res) => {
     }    
 
     const goal = await Goal.create({
-       text: req.body.text 
+       text: req.body.text ,
+       user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -34,6 +38,18 @@ const putGoals = asyncHandler(async (req, res) => {
     if (!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if (!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // check if the goal is the user's goal
+    if (goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error('Not authorized')
     }
     const updatedGoal = await Goal.findByIdAndUpdate(id, req.body ,{new: true})
 
@@ -50,6 +66,19 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if (!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // check if the goal is the user's goal
+    if (goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error('Not authorized')
+    }
+    
     await goal.deleteOne()
 
     res.status(200).json({id: id})
